+++ fixtures/SortableType.php
@@ -42,12 +42,12 @@ final class SortableType implements Type, ElasticsearchType
         return ['type' => 'keyword'];
     }
 
-    public function normalize(mixed $value): null|string|int|bool
+    public function normalize(mixed $value): null|string|int|float|bool
     {
         return $value->toString();
     }
 
-    public function denormalize(null|string|int|bool $value): mixed
+    public function denormalize(null|string|int|float|bool $value): mixed
     {
         if (!\is_string($value)) {
             throw new \LogicException("'$value' is not a string");