+++ src/Definition/Type/IntType.php
@@ -35,12 +35,12 @@ final class IntType implements Type
             ->map(static fn() => new self);
     }
 
-    public function normalize(mixed $value): null|string|int|bool
+    public function normalize(mixed $value): null|string|int|float|bool
     {
         return $value;
     }
 
-    public function denormalize(null|string|int|bool $value): mixed
+    public function denormalize(null|string|int|float|bool $value): mixed
     {
         if (!\is_int($value)) {
             throw new \LogicException("'$value' is not an integer");